@@ -1,15 +1,16 @@
 # If you have OpenSSL installed, we recommend updating
 # the following line to use "https"
-source 'http://rubygems.org'
-ruby '~> 2.3'
+source 'https://rubygems.org'
+ruby '~> 2.6'
 
-gem "slaw", "~>0.3.0"
+gem "slaw", "~>6.0"
 
 gem 'httpclient', '~> 2.7'
 gem 'moneta', '~> 0.8'
-gem "rake", "~> 10.1.0"
+gem "rake", "~> 10.1"
 
-gem "middleman", "~>4.2.1"
+gem "middleman", "~>4.3"
 gem 'hashie', '~> 3.4'
 gem 'middleman-s3_sync', "~> 4.0.3"
-gem 'haml', '~> 4'
+gem 'haml', '~> 5'
+gem 'rack', '2.0.8'
