@@ -23,7 +23,7 @@ set :images_dir, 'img'
 # Build-specific configuration
 configure :build do
   config[:build] = true
-  activate :minify_css
+  #activate :minify_css
   activate :relative_assets
   # Enable cache buster
   activate :asset_hash, ignore: [/favicon.*/]
@@ -47,8 +47,9 @@ Middleman::Extensions.register(:indigo, IndigoMiddlemanExtension)
 activate :indigo
 
 # cache policies
-year = 60*60*24*365
-day = 60*60*24
+hour = 60*60
+day = hour*24
+year = day*365
 
 # we use asset hashing here, so have expiry far in the future
 caching_policy 'text/css', max_age: year
@@ -56,28 +57,34 @@ caching_policy 'application/javascript', max_age: year
 caching_policy 'image/png', max_age: year
 caching_policy 'image/jpeg', max_age: year
 caching_policy 'application/font-woff', max_age: year
-# cache HTML for up to a day
-caching_policy 'text/html', max_age: day
+# cache HTML for up to an hour
+caching_policy 'text/html', max_age: hour
 
 
 # ----------------------------------------------------------------------
 # Website pages
 
 def pages_for_act(act)
-  path = act.frbr_uri.chomp('/')
-
-  for lang in act.languages
-    expression = act.get_expression(lang.code3)
-
-    # full act
-    proxy "#{path}/#{lang.code3}/index.html", "/templates/act/index.html", locals: {act: expression}, ignore: true
-
-    # table of contents
-    proxy "#{path}/#{lang.code3}/contents/index.html", "/templates/act/contents.html", :locals => {act: expression}, :ignore => true
+  if act.stub
+    expressions = [act]
+  else
+    # expressions at the latest date
+    date = act.points_in_time[-1].date
+    expressions = act.expressions.filter { |x| x.expression_date = date }
   end
 
-  # resources
-  proxy "#{path}/resources/index.html", "/templates/act/resources.html", :locals => {act: act}, :ignore => true
+  # pages for the expressions at the latest point in time
+  for expr in expressions
+    # full act
+    path = act.frbr_uri.chomp('/')
+    proxy "#{path}/#{expr.language}/index.html", "/templates/act/index.html", locals: {act: expr}, ignore: true
+  end
+
+  # older expressions
+  for expr in act.expressions
+    path = expr.expression_frbr_uri.chomp('/')
+    proxy "#{path}/index.html", "/templates/act/index.html", locals: {act: expr}, ignore: true
+  end
 end
 
 # municipality microsites
@@ -98,7 +105,7 @@ for lang in region.bylaws.languages
   proxy path, "/templates/region.html", locals: {region: region, language: lang}, ignore: true
 end
 
-region.bylaws.reject(&:stub).each { |bylaw| pages_for_act(bylaw) }
+region.bylaws.each { |bylaw| pages_for_act(bylaw) }
 
 # s3 bucket
 s3_sync_config.bucket = region.bucket
