@@ -29,21 +29,29 @@ ByLawSearch = function() {
     }
 
     $.getJSON('https://srbeugae08.execute-api.eu-west-1.amazonaws.com/default/searchOpenBylaws', params, function(response, textStatus, jqXHR) {
+      var hits = [];
       ladda.stop();
       console.log(response);
 
       response.q = q;
-      var hits = $.map(response.results, function(result) {
-        result.snippet = result._snippet
-          .replace(/^\s*[;:",.()-]+/, '')  // trim leading punctuation
-          .replace(/<b>/g, "<mark>")
-          .replace(/<\/b>/g, "</mark>")
-          .trim();
-
+      response.results.forEach(function(result) {
         result.region = REGIONS[result.locality];
 
-        return result;
+        // only keep those with a region
+        if (result.region) {
+          hits.push(result);
+
+          result.url = (result.region.microsite ? ('https://' + result.region.bucket) : '') +
+            result.frbr_uri + "/" + result.language + "/";
+
+          result.snippet = result._snippet
+            .replace(/^\s*[;:",.()-]+/, '')  // trim leading punctuation
+            .replace(/<b>/g, "<mark>")
+            .replace(/<\/b>/g, "</mark>")
+            .trim();
+        }
       });
+
       response.hits = {hits: hits};
       response.no_region = region_code === "";
       response.regions = [];
@@ -59,23 +67,6 @@ ByLawSearch = function() {
         .show();
 
       $waiting.hide();
-
-      if(getParameterByName('q').match(/(ads|advert|billboard)/)) {
-        $("#help").removeClass("hidden");
-        $(".ads").removeClass("hidden");
-      }
-      if(getParameterByName('q').match(/(animal|dog|cat|pet)/)) {
-        $("#help").removeClass("hidden");
-        $(".animals").removeClass("hidden");
-      }
-      if(getParameterByName('q').match(/(neighbo|contruct|build)/)) {
-        $("#help").removeClass("hidden");
-        $(".neighbours").removeClass("hidden");
-      }
-      if(getParameterByName('q').match(/(nois|music|sound|loud)/)) {
-        $("#help").removeClass("hidden");
-        $(".noise").removeClass("hidden");
-      }
     });
   };
 
